Moving Static supporting classes into Gate.Middleware.StaticFiles namespace.
