+++ src/Gate.Middleware/Gate.Middleware.csproj
@@ -33,21 +33,20 @@
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Web" />
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BodyStream.cs" />
-    <Compile Include="FileBody.cs" />
-    <Compile Include="FileServer.cs" />
+    <Compile Include="StaticFiles\BodyStream.cs" />
+    <Compile Include="StaticFiles\FileBody.cs" />
+    <Compile Include="StaticFiles\FileServer.cs" />
     <Compile Include="MethodOverride.cs" />
-    <Compile Include="Mime.cs" />
+    <Compile Include="StaticFiles\Mime.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="BasicAuth.cs" />
     <Compile Include="Predicate.cs" />
-    <Compile Include="StateMachine.cs" />
+    <Compile Include="StaticFiles\StateMachine.cs" />
     <Compile Include="Static.cs" />
-    <Compile Include="TextBody.cs" />
+    <Compile Include="StaticFiles\TextBody.cs" />
     <Compile Include="Transform.cs" />
     <Compile Include="ContentLength.cs" />
     <Compile Include="Chunked.cs" />
