+++ src/Gate.Middleware/Static.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using Gate.Middleware.StaticFiles;
 using Gate.Owin;
 
 namespace Gate.Middleware
