+++ src/Gate.Middleware/Utils/HttpFormatExtensions.cs
@@ -10,14 +10,5 @@ namespace Gate.Middleware.Utils
             // Sun, 06 Nov 1994 08:49:37 GMT
             return date.ToUniversalTime().ToString("R");
         }
-        public static string Escape(this string uri)
-        {
-            return HttpUtility.UrlEncode(uri);
-        }
-
-        public static string Unescape(this string uri)
-        {
-            return HttpUtility.UrlDecode(uri);
-        }
     }
 }
